backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=1.4
alembic
pydantic>2.0
pydantic-settings
passlib[bcrypt]
python-jose
python-dotenv
email-validator